Distinguish the type of LFSR that is implemented by name
The current implementation implements LFSRs in Fibonacci configuration. LFSRs
in Galois configuration are usually preferred when implemented in hardware,
because in this configuration, all cells can be calculated in parallel. Both
configurations can produce the same sequence of bits, however the seeds (the
initial values loaded into the registers) need to be adjusted to get the same
phase with either of the configurations.

It might be useful to have an implementation of the Galois configuration, in
case a spec uses it and provides seeds for it as well.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>