Add utilities to reverse the order of polynoms and initial state
By default, the generator polynoms and initial state *integers* used in this
library use the order of a shift register that shifts from left to right and
the output falls out of the rightmost register cell.

In theory, you'd have to invert the cell numbering between Fibonacci and Galois
configuration LFSRs. In practice I have not seen this done consistently. Rather
than that, authors seem to prefer to include a sketch of the LFSR and how to
turn the polynom into tap positions and in what order to load the initial word.

These utilities are there to make it easier to convert whatever way a certain
spec may use.

(let ((gp #b101001)
      (is  #b10011))
  (let ((reversed (reverse-gp-and-state gp is))
        (galois (word->bit-lfsr (make-lfsr-stream-galois gp is)))
        (fibona (word->bit-lfsr (make-lfsr-stream-fibonacci (car reversed)
                                                            (cdr reversed)))))
    ;; The galois and fibona streams will generate the same bit
    ;; stream (albeit most likely shifted in phase)...
    ))

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>