Rename feedback-xor to multi-xor
The original intend was for this function to be used in the feedback
calculation in the Fibonacci LFSR configuration. It turned out to be useful
with the phase-selector in C/A stream generation in GPS systems, where the
original name doesn't fit at all.

multi-xor is a more general-name for what this function does, that fits both
use-cases.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>